props passadas pelo app.js
